Feature: setup api and persistence layer

+++ src/Core/Solforb.Orders.Application/Features/Orders/Handlers/Commands/DeleteOrderCommandHandler.cs
@@ -1,13 +1,10 @@
 ﻿using AutoMapper;
 using MediatR;
+using Solforb.Orders.Application.Exceptions;
 using Solforb.Orders.Application.Features.Common;
 using Solforb.Orders.Application.Features.Orders.Requests.Commands;
 using Solforb.Orders.Application.Persistence.Contracts;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using Solforb.Orders.Domain;
 
 namespace Solforb.Orders.Application.Features.Orders.Handlers.Commands
 {
@@ -21,6 +18,9 @@ namespace Solforb.Orders.Application.Features.Orders.Handlers.Commands
 		public override async Task<Unit> Handle(DeleteOrderCommand request, CancellationToken cancellationToken)
 		{
 			var order = await _unitOfWork.OrderRepository.GetByID(request.Id);
+
+			if (order == null)
+				throw new NotFoundException(nameof(Order), request.Id);
 
 			await _unitOfWork.OrderRepository.Delete(order);
